+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.3.1] 2021-02-27
+
+### Fixed
+
+- [PowerShellBuild](https://github.com/psake/PowerShellBuild) dependency is now correctly set to 0.5.0
+
 ## [0.3.0] 2021-02-27
 
 ### Added